The code. RTree implementation and everything else, including gnuplot scripts.
If one wants to run the experiments, unpack the sampleData archive in the same directory as the Makefile.
After that run 'sh runAndPlot.sh' - it will do a couple of checks whether needed programs are in the
PATH variable and exit with an error message if not. Otherwise it will just run through the experiments
and create the plot PDFs.

The default will run the exhaustive algo version just once (it takes a while... ), if an argument is specified
to runAndPlot.sh, then the exhaustive algo version will also be rerun 10 times (warning: if you do that be prepared
to go and make coffee or to play with the pet hamster - it's slooooow).